docs: update Web3ConfigProvider demos
